+++ Modules/Tree/app/Http/Controllers/TreeController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 use Modules\Tree\app\Http\Requests\Tree\StoreRequest;
 use Modules\Tree\app\Models\Land;
@@ -28,7 +29,7 @@ class TreeController extends Controller
      */
     public function index(Request $request, Land $land)
     {
-        $trees = Tree::with('images','thumb_image')->where('land_id', $land->id)->where(function ($query) use ($request) {
+        $trees = Tree::with('images', 'thumb_image')->where('land_id', $land->id)->where(function ($query) use ($request) {
             $query->where('name', 'LIKE', '%' . $request->search . '%');
             $query->orwhere('qr_code', 'LIKE', '%' . $request->search . '%');
             // $query->orwhere('phone', 'LIKE', '%' . $request->term . '%');
@@ -51,6 +52,8 @@ class TreeController extends Controller
      */
     public function store(StoreRequest $request, Land $land): RedirectResponse
     {
+
+
         $tree = Tree::create($request->all());
         if ($tree) {
             $tree->qr_code = $tree->id . Str::random(10);
@@ -61,7 +64,7 @@ class TreeController extends Controller
         foreach ($request->images as $image) {
             $tree->addMedia($image)->toMediaCollection('tree_images');
         }
-        if($request->images_thumb){
+        if ($request->images_thumb) {
             $tree->addMedia($request->images_thumb)->toMediaCollection('tree_thumb');
         }
         return back()->with('success', 'Create successfully');
@@ -86,22 +89,59 @@ class TreeController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Tree $tree): RedirectResponse
+    public function update(Request $request, Tree $tree)
     {
-     
-        $tree->update($request->all());
-        if ($request->hasFile('images')) {
-            $tree->clearMediaCollection('tree_images');
+        $user = Auth::user();
+        if ($tree->product_service_owner) {
+            if ($user->hasPermissionTo('super-admin')) {
+                $tree->update($request->all());
+                if ($request->hasFile('images')) {
+                    $tree->clearMediaCollection('tree_images');
 
-            foreach ($request->images as $image) {
-                $tree->addMedia($image)->toMediaCollection('tree_images');
+                    foreach ($request->images as $image) {
+                        $tree->addMedia($image)->toMediaCollection('tree_images');
+                    }
+                }
+                if ($request->hasFile('images_thumb')) {
+                    $tree->clearMediaCollection('tree_thumb');
+                    $tree->addMedia($request->file('images_thumb'))->toMediaCollection('tree_thumb');
+                }
+            } else {
+                $tree->update([
+                    'name' => $request->name,
+                    "description" => $request->description,
+                    "user_manual" => $request->user_manual,
+                    "terms_policy" => $request->terms_policy,
+
+                ]);
+                if ($request->hasFile('images')) {
+                    $tree->clearMediaCollection('tree_images');
+
+                    foreach ($request->images as $image) {
+                        $tree->addMedia($image)->toMediaCollection('tree_images');
+                    }
+                }
+                if ($request->hasFile('images_thumb')) {
+                    $tree->clearMediaCollection('tree_thumb');
+                    $tree->addMedia($request->file('images_thumb'))->toMediaCollection('tree_thumb');
+                }
+                return back()->with('success', 'Update successfully');
             }
+        } else {
+            $tree->update($request->all());
+            if ($request->hasFile('images')) {
+                $tree->clearMediaCollection('tree_images');
 
+                foreach ($request->images as $image) {
+                    $tree->addMedia($image)->toMediaCollection('tree_images');
+                }
+            }
+            if ($request->hasFile('images_thumb')) {
+                $tree->clearMediaCollection('tree_thumb');
+                $tree->addMedia($request->file('images_thumb'))->toMediaCollection('tree_thumb');
+            }
         }
-        if($request->hasFile('images_thumb')){
-            $tree->clearMediaCollection('tree_thumb');
-            $tree->addMedia($request->file('images_thumb'))->toMediaCollection('tree_thumb');
-        }
+
         return back()->with('success', 'Update successfully');
     }
 
@@ -110,19 +150,23 @@ class TreeController extends Controller
      */
     public function destroy(Tree $tree)
     {
+
+        if ($tree->product_service_owner) {
+            return back()->with('warning', 'Cây cam đã được nhận nuôi bởi khách hàng, Bạn nên xóa hợp đồng trước khi xóa cây');
+        }
         $tree->clearMediaCollection('tree_images');
         $tree->clearMediaCollection('tree_thumb');
         $tree->delete();
         return back()->with('success', 'Delete successfully');
     }
-    public function treeDetail(Request $request,$product_owner_id){
-        $tree = Tree::with('land','thumb_image','images','product_service_owner.customer')->where('product_service_owner_id',$product_owner_id)->first();
+    public function treeDetail(Request $request, $product_owner_id)
+    {
+        $tree = Tree::with('land', 'thumb_image', 'images', 'product_service_owner.customer')->where('product_service_owner_id', $product_owner_id)->first();
         $contact = Contact::find(1);
-        if($tree){
+        if ($tree) {
             // return $tree;
-            return Inertia::render('Modules/Tree/Tree/QrTree', compact('contact','tree'));
+            return Inertia::render('Modules/Tree/Tree/QrTree', compact('contact', 'tree'));
         }
         return response()->json('Không tìm thấy cây nào', 404);
     }
-
 }
